Fixed some warnings from Codacy

+++ expression.c
@@ -218,7 +218,8 @@ enum eExprErr ComputeExpression(const char *expr, BigInteger *ExpressionResult)
       }
       // Move number to compressed stack.
       currentOffset = comprStackOffset[stackIndex];
-      if (currentOffset >= (COMPR_STACK_SIZE - sizeof(BigInteger) / sizeof(limb)))
+      if (currentOffset >= (COMPR_STACK_SIZE - 
+        ((int)sizeof(BigInteger) / (int)sizeof(limb))))
       {
         return EXPR_OUT_OF_MEMORY;
       }
